# in_syslog: do batching to reduce ring buffer pressure under load [v4.0]